Cambios en aside, header y añadir logo
