+++ src/Header.css
@@ -1,0 +1,11 @@
+.header {
+  grid-area: header;
+  padding: 20px;
+  display: flex;
+  gap: 20px;
+  align-items: center;
+  justify-content: space-between;
+  }
+.logo {
+  height: 120px;
+}
