--- src/Logo.js
@@ -1,10 +1,0 @@
-
-
-function Logo() {
-    return (
-        <>
-            <img src='https://www.designevo.com/res/templates/thumb_small/blue-postmark-with-star.webp' alt="logo" width="100px" height="100px"/>
-        </>
-    )
-}
-export default Logo
